# [IMP] stock: optimize move reservation